6. Multi Language Invoice Extractor/requirements.txt
streamlit
google-genai
python-dotenv
langchain
langchain_community
langchain_openai
pillow
